Add navbar module
